requirements.txt
streamlit
langchain
langchain-groq
faiss-cpu
sentence-transformers
pypdf
python-dotenv
langchain-community
google-generativeai
wolframalpha